Add new controller
This adds a new controller. Controllers handle the routing. The index
function is what will show when the name of the controller is used. Then
the functions will represent children routes.

The Welcome function will be called from visiting this URL:
/HelloWorld/Welcome?name=Hairball&exampleNum=27

The WelcomeWithID function will be called from visiting this URL:
/HelloWorld/WelcomeWithID/3?name=Hairball